change commit_time type from Int to Long

+++ git_commit/src/main/kotlin/DTO.kt
@@ -2,7 +2,7 @@ package com.thoughtworks.archguard.git.analyzer
 
 data class CommitLog(
         val id: String,
-        val commit_time: Int,
+        val commit_time: Long,
         val shortMessage: String,
         val committer_name: String,
         val rep_id: Long,
